pyproject.toml
[tool.poetry]
name = "sfxdlebot"
version = "0.0.1"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.2.3"
pyyaml = "^6.0"
validators = "^0.20.0"
cairosvg = "^2.4.2"
drawSvg = "^1.6.1"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
